+++ library/src/main/java/io/sfrei/tracksearch/tracks/SoundCloudTrack.java
@@ -1,21 +1,31 @@
 package io.sfrei.tracksearch.tracks;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import io.sfrei.tracksearch.clients.setup.TrackSource;
 import io.sfrei.tracksearch.tracks.deserializer.SoundCloudTrackDeserializer;
 import io.sfrei.tracksearch.tracks.metadata.SoundCloudTrackInfo;
-import io.sfrei.tracksearch.clients.setup.TrackSource;
 import lombok.Getter;
 import lombok.Setter;
 
-@Getter
-@Setter
+import java.util.function.Function;
+
+
 @JsonDeserialize(using = SoundCloudTrackDeserializer.class)
 public class SoundCloudTrack extends BaseTrack implements Track {
 
+    @Getter
+    @Setter
     private SoundCloudTrackInfo trackInfo;
+
+    @Setter
+    private Function<SoundCloudTrack, String> streamUrlProvider;
 
     public SoundCloudTrack(String title, Long length, String mrl) {
         super(TrackSource.Soundcloud, title, length, mrl);
     }
 
+    @Override
+    public String getStreamUrl() {
+        return streamUrlProvider.apply(this);
+    }
 }