+++ library/src/main/java/io/sfrei/tracksearch/tracks/Track.java
@@ -14,6 +14,8 @@ public interface Track {
 
     String getUrl();
 
+    String getStreamUrl();
+
     boolean equals(Object o);
 
     String toPrettyString();